untap: add new module to make these refactored methods easier to test
